Update keyword_generator.py
The first response!:
Q. What's the key to becoming a good Data Scientist?

A. The Key is to Be Responsive. Responsiveness is in control. The world of business has changed. The focus has changed. This is what has changed. Every business has changed. The Data Science field is about getting people to use the data.

Q. I can't stay on my team in our organization for

LOL!